Address feedback part 2 - moving loki clients
