+++ component/common/loki/client/batch_test.go
