+++ component/common/loki/client/client_test.go
