+++ component/common/loki/client/config_test.go
