+++ component/common/loki/client/fake/client.go
