+++ component/common/loki/client/logger_test.go
