+++ component/common/loki/client/multi.go
