+++ component/common/loki/client/multi_test.go
