+++ component/loki/process/internal/stages/pipeline_test.go
@@ -6,11 +6,10 @@ package stages
 
 import (
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"sync"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
