+++ component/loki/source/cloudflare/internal/cloudflaretarget/target_test.go
@@ -7,12 +7,11 @@ package cloudflaretarget
 import (
 	"context"
 	"errors"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"os"
 	"sort"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/component/common/loki/positions"
