+++ component/loki/source/gcplog/internal/gcplogtarget/pull_target_test.go
@@ -3,11 +3,10 @@ package gcplogtarget
 import (
 	"context"
 	"errors"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"io"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"cloud.google.com/go/pubsub"
 	"github.com/go-kit/log"
