+++ component/loki/source/gcplog/internal/gcplogtarget/push_target_test.go
@@ -2,14 +2,13 @@ package gcplogtarget
 
 import (
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"net/http"
 	"os"
 	"strings"
 	"sync"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/component/common/loki"
