+++ component/loki/source/heroku/internal/herokutarget/target_test.go
@@ -7,6 +7,7 @@ package herokutarget
 import (
 	"flag"
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"net"
 	"net/http"
 	"net/url"
@@ -14,8 +15,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/go-kit/log"
 	"github.com/google/uuid"
