+++ component/loki/source/http/internal/lokipush/pushtarget_test.go
@@ -8,6 +8,7 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"io"
 	"net"
 	"net/http"
@@ -15,8 +16,6 @@ import (
 	"strconv"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/dskit/flagext"
