+++ component/loki/source/internal/kafkatarget/kafkatarget_test.go
@@ -7,11 +7,10 @@ package kafkatarget
 import (
 	"context"
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"sync"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/Shopify/sarama"
 	"github.com/prometheus/common/model"
