+++ component/loki/source/internal/kafkatarget/target_syncer_test.go
@@ -3,11 +3,10 @@ package kafkatarget
 import (
 	"context"
 	"fmt"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 	"reflect"
 	"testing"
 	"time"
-
-	"github.com/grafana/agent/component/loki/internal/fake"
 
 	"github.com/grafana/dskit/flagext"
 	"github.com/prometheus/common/config"
